requirements.txt
attrs==17.4.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pluggy==0.6.0
py==1.5.2
PyJWT==1.5.3
PySocks==1.6.8
pytest==3.4.0
python-dotenv==0.7.1
pytz==2017.3
requests==2.18.4
six==1.11.0
twilio==6.10.2
urllib3==1.22
Werkzeug==0.14.1
